Split up the runtime into multiple files to improve clarity
