+++ src/runtime/mod.rs
@@ -1,4 +1,6 @@
 // modules
+pub mod batches;
+pub mod workpool;
 pub mod runtime;
 
 // flatten
